requirements.txt
pyserial
pyyaml
apscheduler
requests
smbus2
rich